refactor(models): rename ImplicitCommunicationBehavior.InstanceRefs module to InstanceRef (Rule 0007 A)
